requirements.txt
numpy>=1.20.0
tqdm>=4.60.0
requests>=2.25.1
python-docx>=0.8.11
PyPDF2>=2.10.0
pymupdf>=1.19.0
pytesseract>=0.3.8
Pillow>=8.2.0
pandas>=1.3.0
openpyxl>=3.0.7
weaviate-client>=3.0.0
